+++ container-engine-lib/docs/changelog.md
@@ -1,5 +1,13 @@
 # TBD
 
+# 0.2.3
+### Features
+* Verify that, when starting a container with `shouldPublishAllPorts` == `true`, each used port gets exactly one host machine port binding
+
+### Fixes
+* Fixed ports not getting bound when running a container with an `EXPOSE` directive in the image Dockerfile
+* Fixed broken CI
+
 # 0.2.2
 ### Fixes
 * Fixed bug in markdown-link-check property